Move save primary term logic to the watcher
The WPSEO_Primary_Term_Admin is loaded conditionally and is responsible for showing the UI regarding the primary terms.

The watcher should be responsible for handing changes.
Moved the code accordingly.
Untested.
